feat(core): Add jmh examples
Add jmh examples.

BREAKING CHANGE: Add jmh examples.
